rename utils folders
